+++ implementation/src/main/java/com/avioconsulting/mule/testing/muleinterfaces/wrappers/InvokeExceptionWrapper.groovy
@@ -11,7 +11,8 @@ class InvokeExceptionWrapper extends Exception {
      * use this method
      * @return
      */
-    String getMessageAsString() {
-        'no way jose'
+    MessageWrapper getMuleMessage() {
+        assert cause.getClass().getName().contains('MessagingException')
+        new MessageWrapperImpl(cause.muleMessage)
     }
 }