+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Consistent logging
 * Standardize closure currying for XML/REST and SOAP to allow getting connector info in the mock closure
 * Add ability to get access to SOAP headers when mocking SOAP calls
+* Wrap exceptions thrown from flow invocations to make it easier to assert message contents
 
 # 2.0.6
 * Add support for `target` and ensure mocked connectors that have it configured have mocked output sent to the appropriate flow variable. NOTE: `targetValue` is not supported yet